move files around; remove max_memory_mb
