+++ openhands/tools/execute_bash/impl.py
@@ -3,7 +3,7 @@ from openhands.tools.execute_bash.definition import (
     ExecuteBashAction,
     ExecuteBashObservation,
 )
-from openhands.tools.execute_bash.sessions.factory import create_terminal_session
+from openhands.tools.execute_bash.terminal.factory import create_terminal_session
 
 
 class BashExecutor(ToolExecutor):
@@ -11,7 +11,6 @@ class BashExecutor(ToolExecutor):
         self,
         working_dir: str,
         username: str | None = None,
-        max_memory_mb: int | None = None,
         no_change_timeout_seconds: int | None = None,
         session_type: str | None = None,
     ):
@@ -20,7 +19,6 @@ class BashExecutor(ToolExecutor):
         Args:
             working_dir: Working directory for bash commands
             username: Optional username for the bash session
-            max_memory_mb: Optional memory limit in MB
             no_change_timeout_seconds: Timeout for no output change
             session_type: Force a specific session type:
                          ('tmux', 'subprocess', 'powershell').
@@ -29,7 +27,6 @@ class BashExecutor(ToolExecutor):
         self.session = create_terminal_session(
             work_dir=working_dir,
             username=username,
-            max_memory_mb=max_memory_mb,
             no_change_timeout_seconds=no_change_timeout_seconds,
             session_type=session_type,
         )
