--- openhands/tools/execute_bash/sessions/__init__.py
@@ -1,33 +1,0 @@
-# Terminal session implementations - New Unified Architecture
-
-# Factory for creating sessions
-from openhands.tools.execute_bash.sessions.factory import create_terminal_session
-from openhands.tools.execute_bash.sessions.powershell_terminal import PowershellTerminal
-from openhands.tools.execute_bash.sessions.subprocess_terminal import SubprocessTerminal
-
-# Terminal interface and backends
-from openhands.tools.execute_bash.sessions.terminal_interface import TerminalInterface
-from openhands.tools.execute_bash.sessions.tmux_terminal import TmuxTerminal
-
-# Unified session controller and base classes
-from openhands.tools.execute_bash.sessions.unified_session import (
-    TerminalCommandStatus,
-    TerminalSession,
-    UnifiedBashSession,
-)
-
-
-__all__ = [
-    # === Factory ===
-    "create_terminal_session",
-    # === Base Classes ===
-    "TerminalSession",
-    "TerminalCommandStatus",
-    # === Terminal Interface and Backends ===
-    "TerminalInterface",
-    "TmuxTerminal",
-    "SubprocessTerminal",
-    "PowershellTerminal",
-    # === Unified Session Controller ===
-    "UnifiedBashSession",
-]
