+++ openhands/tools/execute_bash/terminal/__init__.py
@@ -1,0 +1,24 @@
+from openhands.tools.execute_bash.terminal.factory import create_terminal_session
+from openhands.tools.execute_bash.terminal.interface import (
+    TerminalInterface,
+    TerminalSessionBase,
+)
+from openhands.tools.execute_bash.terminal.powershell_terminal import PowershellTerminal
+from openhands.tools.execute_bash.terminal.subprocess_terminal import SubprocessTerminal
+from openhands.tools.execute_bash.terminal.terminal_session import (
+    TerminalCommandStatus,
+    TerminalSession,
+)
+from openhands.tools.execute_bash.terminal.tmux_terminal import TmuxTerminal
+
+
+__all__ = [
+    "TerminalInterface",
+    "TerminalSessionBase",
+    "TmuxTerminal",
+    "SubprocessTerminal",
+    "PowershellTerminal",
+    "TerminalSession",
+    "TerminalCommandStatus",
+    "create_terminal_session",
+]
