+++ openhands/tools/execute_bash/terminal/factory.py
@@ -4,7 +4,7 @@ import platform
 import subprocess
 
 from openhands.sdk.logger import get_logger
-from openhands.tools.execute_bash.sessions.unified_session import TerminalSession
+from openhands.tools.execute_bash.terminal.terminal_session import TerminalSession
 
 
 logger = get_logger(__name__)
@@ -48,7 +48,6 @@ def _is_powershell_available() -> bool:
 def create_terminal_session(
     work_dir: str,
     username: str | None = None,
-    max_memory_mb: int | None = None,
     no_change_timeout_seconds: int | None = None,
     session_type: str | None = None,
 ) -> TerminalSession:
@@ -57,7 +56,6 @@ def create_terminal_session(
     Args:
         work_dir: Working directory for the session
         username: Optional username for the session
-        max_memory_mb: Optional memory limit in MB
         no_change_timeout_seconds: Timeout for no output change
         session_type: Force a specific session type ('tmux', 'subprocess', 'powershell')
                      If None, auto-detect based on system capabilities
@@ -68,36 +66,36 @@ def create_terminal_session(
     Raises:
         RuntimeError: If the requested session type is not available
     """
-    from openhands.tools.execute_bash.sessions.unified_session import UnifiedBashSession
+    from openhands.tools.execute_bash.terminal.terminal_session import TerminalSession
 
     if session_type:
         # Force specific session type
         if session_type == "tmux":
             if not _is_tmux_available():
                 raise RuntimeError("Tmux is not available on this system")
-            from openhands.tools.execute_bash.sessions.tmux_terminal import TmuxTerminal
+            from openhands.tools.execute_bash.terminal.tmux_terminal import TmuxTerminal
 
             logger.info("Using forced TmuxTerminal")
-            terminal = TmuxTerminal(work_dir, username, max_memory_mb)
-            return UnifiedBashSession(terminal, no_change_timeout_seconds)
+            terminal = TmuxTerminal(work_dir, username)
+            return TerminalSession(terminal, no_change_timeout_seconds)
         elif session_type == "subprocess":
-            from openhands.tools.execute_bash.sessions.subprocess_terminal import (
+            from openhands.tools.execute_bash.terminal.subprocess_terminal import (
                 SubprocessTerminal,
             )
 
             logger.info("Using forced SubprocessTerminal")
-            terminal = SubprocessTerminal(work_dir, username, max_memory_mb)
-            return UnifiedBashSession(terminal, no_change_timeout_seconds)
+            terminal = SubprocessTerminal(work_dir, username)
+            return TerminalSession(terminal, no_change_timeout_seconds)
         elif session_type == "powershell":
             if not _is_powershell_available():
                 raise RuntimeError("PowerShell is not available on this system")
-            from openhands.tools.execute_bash.sessions.powershell_terminal import (
+            from openhands.tools.execute_bash.terminal.powershell_terminal import (
                 PowershellTerminal,
             )
 
             logger.info("Using forced PowershellTerminal")
-            terminal = PowershellTerminal(work_dir, username, max_memory_mb)
-            return UnifiedBashSession(terminal, no_change_timeout_seconds)
+            terminal = PowershellTerminal(work_dir, username)
+            return TerminalSession(terminal, no_change_timeout_seconds)
         else:
             raise ValueError(f"Unknown session type: {session_type}")
 
@@ -107,15 +105,15 @@ def create_terminal_session(
     if system == "Windows":
         # On Windows, prefer PowerShell if available, otherwise use subprocess
         if _is_powershell_available():
-            from openhands.tools.execute_bash.sessions.powershell_terminal import (
+            from openhands.tools.execute_bash.terminal.powershell_terminal import (
                 PowershellTerminal,
             )
 
             logger.info("Auto-detected: Using PowershellTerminal on Windows")
-            terminal = PowershellTerminal(work_dir, username, max_memory_mb)
-            return UnifiedBashSession(terminal, no_change_timeout_seconds)
+            terminal = PowershellTerminal(work_dir, username)
+            return TerminalSession(terminal, no_change_timeout_seconds)
         else:
-            from openhands.tools.execute_bash.sessions.subprocess_terminal import (
+            from openhands.tools.execute_bash.terminal.subprocess_terminal import (
                 SubprocessTerminal,
             )
 
@@ -123,21 +121,21 @@ def create_terminal_session(
                 "Auto-detected: Using SubprocessTerminal on Windows "
                 "(PowerShell not available)"
             )
-            terminal = SubprocessTerminal(work_dir, username, max_memory_mb)
-            return UnifiedBashSession(terminal, no_change_timeout_seconds)
+            terminal = SubprocessTerminal(work_dir, username)
+            return TerminalSession(terminal, no_change_timeout_seconds)
     else:
         # On Unix-like systems, prefer tmux if available, otherwise use subprocess
         if _is_tmux_available():
-            from openhands.tools.execute_bash.sessions.tmux_terminal import TmuxTerminal
+            from openhands.tools.execute_bash.terminal.tmux_terminal import TmuxTerminal
 
             logger.info("Auto-detected: Using TmuxTerminal (tmux available)")
-            terminal = TmuxTerminal(work_dir, username, max_memory_mb)
-            return UnifiedBashSession(terminal, no_change_timeout_seconds)
+            terminal = TmuxTerminal(work_dir, username)
+            return TerminalSession(terminal, no_change_timeout_seconds)
         else:
-            from openhands.tools.execute_bash.sessions.subprocess_terminal import (
+            from openhands.tools.execute_bash.terminal.subprocess_terminal import (
                 SubprocessTerminal,
             )
 
             logger.info("Auto-detected: Using SubprocessTerminal (tmux not available)")
-            terminal = SubprocessTerminal(work_dir, username, max_memory_mb)
-            return UnifiedBashSession(terminal, no_change_timeout_seconds)
+            terminal = SubprocessTerminal(work_dir, username)
+            return TerminalSession(terminal, no_change_timeout_seconds)
