+++ tests/tools/execute_bash/conftest.py
@@ -4,7 +4,7 @@ import tempfile
 
 from openhands.sdk.logger import get_logger
 from openhands.tools.execute_bash.constants import TIMEOUT_MESSAGE_TEMPLATE
-from openhands.tools.execute_bash.sessions import create_terminal_session
+from openhands.tools.execute_bash.terminal import create_terminal_session
 
 
 logger = get_logger(__name__)
